# web/openapi.yaml
openapi: 3.0.3
info:
  title: work-tracker API
  description: Automatically generated OpenAPI spec from Postman collection "work-tracker"
  version: 1.0.0
servers:
  - url: http://127.0.0.1:5000
    description: Default local server
paths:
...
  /api/categories:
    post:
      summary: Add a task category (admin)
      operationId: createCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkOptionRequest'
      responses:
        '200':
          $ref: '#/components/responses/WorkOptionOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/categories/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    put:
      summary: Rename or (de)activate a task category (admin)
      operationId: updateCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkOptionRequest'
      responses:
        '200':
          $ref: '#/components/responses/WorkOptionOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a task category (admin)
      description: Work records and todos using it get an empty categoryId. `cleared` is how many were changed.
      operationId: deleteCategory
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteWorkOptionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/projects:
    post:
      summary: Add a project (admin)
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkOptionRequest'
      responses:
        '200':
          $ref: '#/components/responses/WorkOptionOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/projects/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    put:
      summary: Rename or (de)activate a project (admin)
      operationId: updateProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkOptionRequest'
      responses:
        '200':
          $ref: '#/components/responses/WorkOptionOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a project (admin)
      description: Work records and todos using it get an empty projectId. `cleared` is how many were changed.
      operationId: deleteProject
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteWorkOptionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/settings/work:
    put:
      summary: Update the work table setting (admin)
      operationId: updateWorkSetting
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkSettingRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkSettingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    WorkOptionOk:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WorkOptionResponse'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
  schemas:
...
    CategoryColor:
      type: string
      description: Palette color name of a task category. Projects have no color. New categories get the least used color.
      enum:
        - blue
        - sky
        - teal
        - green
        - lime
        - amber
        - orange
        - red
        - pink
        - purple
    WorkOption:
      type: object
      required:
        - id
        - name
        - active
      properties:
        id:
          type: string
        name:
          type: string
        active:
          type: boolean
          description: Inactive options are hidden from new entries but still name old ones.
        color:
          $ref: '#/components/schemas/CategoryColor'
...
    CreateWorkOptionRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
    UpdateWorkOptionRequest:
      type: object
      properties:
        name:
          type: string
        active:
          type: boolean
        color:
          $ref: '#/components/schemas/CategoryColor'
    WorkOptionResponse:
      type: object
      required:
        - message
        - option
      properties:
        message:
          type: string
        option:
          $ref: '#/components/schemas/WorkOption'
    DeleteWorkOptionResponse:
      type: object
      required:
        - message
        - cleared
      properties:
        message:
          type: string
        cleared:
          type: integer
          description: Number of work records and todos that no longer reference the deleted option.
    UpdateWorkSettingRequest:
      type: object
      required:
        - allowViewAll
      properties:
        allowViewAll:
          type: boolean
    WorkSettingResponse:
      type: object
      required:
        - message
        - allowViewAll
      properties:
        message:
          type: string
        allowViewAll:
          type: boolean
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string